add PPMUNet

+++ basicsr/utils/img_util.py
@@ -94,7 +94,7 @@ def tensor2img(tensor, rgb2bgr=True, out_type=np.uint8, min_max=(0, 1)):
     return result
 
 
-def tensor2npy(tensor, out_type=np.float):
+def tensor2npy(tensor, out_type=np.float32):
     """ Convert torch Tensors into image numpy arrays.
     After clamping to [min, max], values will be normalized to [0, 1].
